Improving code readability
