--- public/country.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-  <style>
-    ul {
-      list-style: none;
-      margin: 0;
-      padding: 0;
-      font-family: sans-serif;
-    }
-
-    a {
-      clear: both;
-      margin-top: 40px;
-      text-decoration: none;
-    }
-
-    img {
-      width: 100px;
-      display: block;
-      margin-right: 20px;
-    }
-
-    p {
-      margin: 0;
-    }
-  </style>
-</head>
-
-<body>
-  <a href="index.html">Back</a>
-  <ul id="countries-list"></ul>
-</body>
-<script src="bundle.js"></script>
-
-</html>
