+++ public/css/styles.css
@@ -381,7 +381,7 @@ form {
   position: relative; }
   form .btn-search {
     position: absolute;
-    top: 50%;
+    top: 25%;
     left: 5%; }
 
 input[type=text] {
@@ -389,7 +389,6 @@ input[type=text] {
   width: 100%;
   border-radius: 8px;
   border: none;
-  margin-top: 1.875em;
   padding: 0.9375em;
   padding-left: 20%;
   vertical-align: top;
@@ -403,11 +402,11 @@ button[type=submit] {
   border: 0;
   background-color: white;
   color: #858585;
-  top: -1.875em;
   cursor: pointer;
   outline: none; }
 
 select {
+  width: 50%;
   border-radius: 8px;
   border: none;
   background-color: white;
@@ -426,7 +425,8 @@ p {
   margin-bottom: 8px; }
 
 ul {
-  padding-left: 0; }
+  padding-left: 0;
+  margin: 0; }
 
 li {
   list-style: none; }
@@ -438,6 +438,12 @@ li {
   padding-right: 1.875em;
   margin-left: 0;
   margin-right: 0; }
+
+/* Grid elements */
+.grid-page {
+  display: grid;
+  grid-template-rows: repeat(3, auto);
+  grid-gap: 2em; }
 
 /* Modules imports */
 /* Header component styles */
@@ -452,19 +458,22 @@ li {
     align-items: center; }
 
 /* Box component styles */
+.grid-box {
+  display: grid;
+  grid-gap: 2em; }
+
 .box {
-  margin-bottom: 1.875em;
   margin-left: auto;
   margin-right: auto;
-  max-width: 23.438em;
+  max-width: 28.125em;
   background-color: white;
-  border-radius: 5px;
+  border-radius: 10px;
   -webkit-box-shadow: 0px 5px 16px -9px rgba(0, 0, 0, 0.75);
   -moz-box-shadow: 0px 5px 16px -9px rgba(0, 0, 0, 0.75);
   box-shadow: 0px 5px 16px -9px rgba(0, 0, 0, 0.75);
   overflow: hidden; }
   .box img {
-    max-width: 100%; }
+    width: 100%; }
   .box .box-info {
     padding: 1.875em; }
   .box label {
