+++ src/scss/layouts/layout.scss
@@ -6,3 +6,10 @@
     margin-left: 0;
     margin-right: 0;
 }
+
+/* Grid elements */
+.grid-page {
+    display: grid;
+    grid-template-rows: repeat(3, auto);
+    grid-gap: 2em;
+}
