+++ src/scss/modules/box.scss
@@ -1,15 +1,18 @@
 /* Box component styles */
+.grid-box {
+    display: grid;
+    grid-gap: 2em;
+}
 .box {
-    margin-bottom: $spacing-default;
     margin-left: auto;
     margin-right: auto;
-    max-width: 23.438em;
+    max-width: 28.125em;
     background-color: $white;
-    border-radius: 5px;
+    border-radius: 10px;
     @include createShadow();
     overflow: hidden;
     img {
-        max-width: 100%;
+        width: 100%;
     }
     .box-info {
         padding: $spacing-default;
